Rename toggle scene mode component
