--- src/components/widgets/toggle-scene/toggle-scene-component.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-const ToggleSceneComponent = props => {
-  const { handleSceneModeChange, sceneMode } = props;
-  return (
-    <button
-      onClick={handleSceneModeChange}
-      style={{ color: 'white', top: 0, position: 'absolute' }}
-    >
-      TOGGLE SCENE MODE
-    </button>
-  );
-}
-
-export default ToggleSceneComponent;
